requirements.txt
flask
requests
python-dotenv
python-telegram-bot==13.15
fastapi
uvicorn
